+++ src/System.IO.Pipelines/System/IO/Pipelines/PipeWriter.cs
@@ -3,6 +3,7 @@
 
 using System.Buffers;
 using System.Threading;
+using System.Threading.Tasks;
 
 namespace System.IO.Pipelines
 {
@@ -11,6 +12,8 @@ namespace System.IO.Pipelines
     /// </summary>
     public abstract class PipeWriter: IOutput
     {
+        private static readonly Task _completedTask = Task.FromResult(0);
+
         /// <summary>
         /// Marks the pipeline as being complete, meaning no more items will be written to it.
         /// </summary>
@@ -33,5 +36,24 @@ namespace System.IO.Pipelines
         public abstract void Advance(int bytes);
         public abstract Memory<byte> GetMemory(int minimumLength = 0);
         public abstract Span<byte> GetSpan(int minimumLength = 0);
+
+        public virtual Task WriteAsync(ReadOnlyMemory<byte> source)
+        {
+            this.Write(source.Span);
+
+            var awaitable = FlushAsync();
+            if (awaitable.IsCompleted)
+            {
+                awaitable.GetResult();
+                return _completedTask;
+            }
+
+            return FlushAsyncAwaited(awaitable);
+        }
+
+        private static async Task FlushAsyncAwaited(ValueAwaiter<FlushResult> awaitable)
+        {
+            await awaitable;
+        }
     }
 }
